+++ R/bin_time.R
@@ -123,6 +123,8 @@ bin_time <- function(
   fun = dunif,
   ...
 ) {
+  ensure_args_are_named(exceptions = "occdf")
+
   check_data_frame(occdf)
   check_data_frame(bins)
 
@@ -134,8 +136,10 @@ bin_time <- function(
 
   occdf_min_ma_vals <- occdf[[min_ma]]
   occdf_max_ma_vals <- occdf[[max_ma]]
-  check_numeric(occdf_min_ma_vals, allow_na = FALSE, arg = "min_ma")
-  check_numeric(occdf_max_ma_vals, allow_na = FALSE, arg = "max_ma")
+  check_na(occdf, min_ma)
+  check_na(occdf, max_ma)
+  check_class(occdf, min_ma, "numeric")
+  check_class(occdf, max_ma, "numeric")
   check_min_lower_than_max(occdf, min_ma, max_ma)
 
   method <- rlang::arg_match(
@@ -143,7 +147,7 @@ bin_time <- function(
     values = c("mid", "majority", "all", "random", "point")
   )
 
-  check_numeric(reps, required_length = 1)
+  rlang::check_number_whole(reps, min = 1)
 
   bins_min_ma_vals <- bins[[min_ma]]
   bins_max_ma_vals <- bins[[max_ma]]