+++ R/tax_expand_time.R
@@ -63,11 +63,14 @@ tax_expand_time <- function(
   rank = "stage",
   ext_orig = TRUE
 ) {
+  ensure_args_are_named(exceptions = "taxdf")
   check_data_frame(taxdf)
 
   check_column_presence(taxdf, max_ma)
   check_column_presence(taxdf, min_ma)
 
+  check_class(taxdf, min_ma, "numeric")
+  check_class(taxdf, max_ma, "numeric")
   check_range(taxdf, min_ma, 0, Inf)
   check_range(taxdf, max_ma, 0, Inf)