+++ NEWS.md
@@ -1,11 +1,21 @@
 # palaeoverse (development version)
 
+This version contains many breaking changes. See the 2.0 migration guide for
+guidance to update existing code.
+
+## Breaking changes
+
 * `palaeoverse` requires R >= 4.1.0 (#181).
 
 ## Breaking changes
 
 * In `tax_check()`, the argument `verbose` is removed and `tax_check()` always returns
-  a `data.frame` with attributes `non_letter_name` and `non_letter_group` (#342)
+  a `data.frame` with attributes `non_letter_name` and `non_letter_group` (#342).
+  
+* All functions now require arguments to be fully named, except for the first
+  argument in some cases (#216).
+
+* In `lat_bins_area()`, the argument `n` is deprecated and replaced by `n_bins` (#349).
 
 ## Bug fixes